[CDS-50913] reverting polling classes
