+++ 120-ng-manager/src/main/java/io/harness/ng/webhook/polling/PollingResponseHandler.java
@@ -48,17 +48,13 @@ import io.harness.logging.AutoLogContext;
 import io.harness.logging.CommandExecutionStatus;
 import io.harness.logging.NgPollingAutoLogContext;
 import io.harness.polling.artifact.ArtifactCollectionUtilsNg;
-import io.harness.polling.bean.ArtifactInfo;
-import io.harness.polling.bean.ArtifactPolledResponse;
-import io.harness.polling.bean.GitPollingPolledResponse;
-import io.harness.polling.bean.HelmChartManifestInfo;
-import io.harness.polling.bean.ManifestInfo;
-import io.harness.polling.bean.ManifestPolledResponse;
 import io.harness.polling.bean.PolledResponseResult;
 import io.harness.polling.bean.PolledResponseResult.PolledResponseResultBuilder;
 import io.harness.polling.bean.PollingDocument;
 import io.harness.polling.bean.artifact.AMIArtifactInfo;
 import io.harness.polling.bean.artifact.AcrArtifactInfo;
+import io.harness.polling.bean.artifact.ArtifactInfo;
+import io.harness.polling.bean.artifact.ArtifactPolledResponse;
 import io.harness.polling.bean.artifact.ArtifactoryRegistryArtifactInfo;
 import io.harness.polling.bean.artifact.AzureArtifactsInfo;
 import io.harness.polling.bean.artifact.CustomArtifactInfo;
@@ -71,6 +67,10 @@ import io.harness.polling.bean.artifact.JenkinsArtifactInfo;
 import io.harness.polling.bean.artifact.Nexus2RegistryArtifactInfo;
 import io.harness.polling.bean.artifact.NexusRegistryArtifactInfo;
 import io.harness.polling.bean.artifact.S3ArtifactInfo;
+import io.harness.polling.bean.gitpolling.GitPollingPolledResponse;
+import io.harness.polling.bean.manifest.HelmChartManifestInfo;
+import io.harness.polling.bean.manifest.ManifestInfo;
+import io.harness.polling.bean.manifest.ManifestPolledResponse;
 import io.harness.polling.contracts.BuildInfo;
 import io.harness.polling.contracts.PollingResponse;
 import io.harness.polling.service.intfc.PollingPerpetualTaskService;
