+++ 120-ng-manager/src/test/java/io/harness/ng/webhook/polling/PollingResponseHandlerTest.java
@@ -59,17 +59,12 @@ import io.harness.delegate.task.artifacts.response.ArtifactDelegateResponse;
 import io.harness.gitpolling.github.GitPollingWebhookData;
 import io.harness.k8s.model.HelmVersion;
 import io.harness.logging.CommandExecutionStatus;
-import io.harness.polling.bean.ArtifactPolledResponse;
-import io.harness.polling.bean.GitHubPollingInfo;
-import io.harness.polling.bean.GitPollingInfo;
-import io.harness.polling.bean.GitPollingPolledResponse;
-import io.harness.polling.bean.HelmChartManifestInfo;
-import io.harness.polling.bean.ManifestPolledResponse;
 import io.harness.polling.bean.PolledResponse;
 import io.harness.polling.bean.PollingDocument;
 import io.harness.polling.bean.PollingInfo;
 import io.harness.polling.bean.PollingType;
 import io.harness.polling.bean.artifact.AcrArtifactInfo;
+import io.harness.polling.bean.artifact.ArtifactPolledResponse;
 import io.harness.polling.bean.artifact.ArtifactoryRegistryArtifactInfo;
 import io.harness.polling.bean.artifact.DockerHubArtifactInfo;
 import io.harness.polling.bean.artifact.EcrArtifactInfo;
@@ -77,6 +72,11 @@ import io.harness.polling.bean.artifact.GARArtifactInfo;
 import io.harness.polling.bean.artifact.GcrArtifactInfo;
 import io.harness.polling.bean.artifact.GithubPackagesArtifactInfo;
 import io.harness.polling.bean.artifact.NexusRegistryArtifactInfo;
+import io.harness.polling.bean.gitpolling.GitHubPollingInfo;
+import io.harness.polling.bean.gitpolling.GitPollingInfo;
+import io.harness.polling.bean.gitpolling.GitPollingPolledResponse;
+import io.harness.polling.bean.manifest.HelmChartManifestInfo;
+import io.harness.polling.bean.manifest.ManifestPolledResponse;
 import io.harness.polling.contracts.PollingResponse;
 import io.harness.polling.contracts.Type;
 import io.harness.polling.service.intfc.PollingPerpetualTaskService;
