+++ 125-cd-nextgen/BUILD.bazel
@@ -19,7 +19,26 @@ java_library(
         "//125-cd-nextgen/modules/azure/core/src/main/java/io/harness/cdng/azure/webapp/steps:module",
         "//125-cd-nextgen/modules/common/core/src/main/java/io/harness/beans/common:module",
         "//125-cd-nextgen/modules/common/core/src/main/java/io/harness/cdng:module",
-        "//125-cd-nextgen/modules/common/core/src/main/java/io/harness/cdng/artifact/bean/artifactsource:module",
+        "//125-cd-nextgen/modules/common/core/src/main/java/io/harness/cdng/artifact/mappers:module",
+        "//125-cd-nextgen/modules/common/core/src/main/java/io/harness/cdng/artifact/resources/acr/service:module",
+        "//125-cd-nextgen/modules/common/core/src/main/java/io/harness/cdng/artifact/resources/ami:module",
+        "//125-cd-nextgen/modules/common/core/src/main/java/io/harness/cdng/artifact/resources/artifactory/service:module",
+        "//125-cd-nextgen/modules/common/core/src/main/java/io/harness/cdng/artifact/resources/azureartifacts:module",
+        "//125-cd-nextgen/modules/common/core/src/main/java/io/harness/cdng/artifact/resources/custom:module",
+        "//125-cd-nextgen/modules/common/core/src/main/java/io/harness/cdng/artifact/resources/docker/service:module",
+        "//125-cd-nextgen/modules/common/core/src/main/java/io/harness/cdng/artifact/resources/ecr/service:module",
+        "//125-cd-nextgen/modules/common/core/src/main/java/io/harness/cdng/artifact/resources/gcr/service:module",
+        "//125-cd-nextgen/modules/common/core/src/main/java/io/harness/cdng/artifact/resources/githubpackages/service:module",
+        "//125-cd-nextgen/modules/common/core/src/main/java/io/harness/cdng/artifact/resources/googleartifactregistry/service:module",
+        "//125-cd-nextgen/modules/common/core/src/main/java/io/harness/cdng/artifact/resources/googlecloudstorage/service:module",
+        "//125-cd-nextgen/modules/common/core/src/main/java/io/harness/cdng/artifact/resources/jenkins/service:module",
+        "//125-cd-nextgen/modules/common/core/src/main/java/io/harness/cdng/artifact/resources/nexus/service:module",
+        "//125-cd-nextgen/modules/common/core/src/main/java/io/harness/cdng/artifact/service:module",
+        "//125-cd-nextgen/modules/common/core/src/main/java/io/harness/cdng/artifact/service/impl:module",
+        "//125-cd-nextgen/modules/common/core/src/main/java/io/harness/cdng/artifact/steps:module",
+        "//125-cd-nextgen/modules/common/core/src/main/java/io/harness/cdng/artifact/steps/beans:module",
+        "//125-cd-nextgen/modules/common/core/src/main/java/io/harness/cdng/artifact/steps/constants:module",
+        "//125-cd-nextgen/modules/common/core/src/main/java/io/harness/cdng/configfile:module",
         "//125-cd-nextgen/modules/common/core/src/main/java/io/harness/cdng/configfile/steps:module",
         "//125-cd-nextgen/modules/common/core/src/main/java/io/harness/cdng/creator/plan:module",
         "//125-cd-nextgen/modules/common/core/src/main/java/io/harness/cdng/creator/plan/environment:module",
@@ -59,14 +78,11 @@ java_library(
         "//125-cd-nextgen/modules/common/core/src/main/java/io/harness/cdng/service/steps:module",
         "//125-cd-nextgen/modules/common/core/src/main/java/io/harness/cdng/ssh/output:module",
         "//125-cd-nextgen/modules/common/core/src/main/java/io/harness/ng:module",
-        "//125-cd-nextgen/modules/common/core/src/main/java/io/harness/polling/bean:module",
+        "//125-cd-nextgen/modules/common/core/src/main/java/io/harness/ng/core/envGroup:module",
         "//125-cd-nextgen/modules/common/core/src/main/java/io/harness/rbac:module",
-        "//125-cd-nextgen/modules/common/core/src/main/java/io/harness/repositories/artifact:module",
         "//125-cd-nextgen/modules/common/core/src/main/java/io/harness/repositories/envGroup:module",
         "//125-cd-nextgen/modules/common/core/src/main/java/io/harness/repositories/executions:module",
-        "//125-cd-nextgen/modules/common/core/src/main/java/io/harness/repositories/polling:module",
-        "//125-cd-nextgen/modules/common/core/src/main/java/io/harness/repositories/telemetry:module",
-        "//125-cd-nextgen/modules/common/core/src/main/java/io/harness/telemetry/beans:module",
+        "//125-cd-nextgen/modules/common/core/src/main/java/io/harness/utils:module",
         "//125-cd-nextgen/modules/gitops/core/src/main/java/io/harness/cdng/gitops/steps:module",
         "//126-instance:module",
         "//127-cd-nextgen-entities:module",
