+++ 125-cd-nextgen/modules/common/core/src/main/java/io/harness/cdng/execution/BUILD.bazel
@@ -14,12 +14,12 @@ java_library(
         "//:lombok",
         "//127-cd-nextgen-entities:module",
         "//884-pms-commons:module",
+        "//960-persistence:module",
         "//960-persistence/src/main/java/io/harness/annotation:module",
-        "//960-persistence/src/main/java/io/harness/mongo/index:module",
         "//960-persistence/src/main/java/io/harness/ng:module",
         "//960-persistence/src/main/java/io/harness/persistence/api:module",
         "//970-ng-commons:module",
-        "//979-recaster/src/main/java/io/harness/annotation:module",
+        "//979-recaster:module",
         "//980-commons/src/main/java/io/harness/annotations:module",
         "//980-commons/src/main/java/io/harness/utils:module",
         "//999-annotations/src/main/java/io/harness/annotations/dev:module",
