--- 125-cd-nextgen/modules/common/core/src/main/java/io/harness/polling/bean/BUILD.bazel
@@ -1,35 +1,0 @@
-# Copyright 2023 Harness Inc. All rights reserved.
-# Use of this source code is governed by the PolyForm Free Trial 1.0.0 license
-# that can be found in the licenses directory at the root of this repository, also available at
-# https://polyformproject.org/wp-content/uploads/2020/05/PolyForm-Free-Trial-1.0.0.txt.
-
-load("@rules_java//java:defs.bzl", "java_library")
-load("//:tools/bazel/macros.bzl", "run_analysis_per_module")
-
-java_library(
-    name = "module",
-    srcs = glob(["*.java"]),
-    visibility = ["//visibility:public"],
-    deps = [
-        "//:lombok",
-        "//127-cd-nextgen-entities:module",
-        "//950-delegate-tasks-beans:module",
-        "//957-cg-beans/src/main/java/software/wings/beans/trigger:module",
-        "//960-persistence/src/main/java/io/harness/annotation:module",
-        "//960-persistence/src/main/java/io/harness/mongo/index:module",
-        "//960-persistence/src/main/java/io/harness/ng:module",
-        "//960-persistence/src/main/java/io/harness/persistence/api:module",
-        "//970-api-services-beans:module",
-        "//970-ng-commons:module",
-        "//980-commons/src/main/java/io/harness/annotations:module",
-        "//999-annotations/src/main/java/io/harness/annotations/dev:module",
-        "@maven//:com_fasterxml_jackson_core_jackson_annotations",
-        "@maven//:com_google_guava_guava",
-        "@maven//:dev_morphia_morphia_core",
-        "@maven//:javax_validation_validation_api",
-        "@maven//:org_springframework_data_spring_data_commons",
-        "@maven//:org_springframework_data_spring_data_mongodb",
-    ],
-)
-
-run_analysis_per_module()
