+++ 125-cd-nextgen/modules/common/core/src/main/java/io/harness/repositories/executions/CDAccountExecutionMetadataRepository.java
@@ -19,6 +19,6 @@ import org.springframework.data.repository.PagingAndSortingRepository;
 @HarnessRepo
 @OwnedBy(PIPELINE)
 public interface CDAccountExecutionMetadataRepository
-    extends PagingAndSortingRepository<CDAccountExecutionMetadata, String>, CDAccountExecutionMetadataRepositoryCustom {
+    extends PagingAndSortingRepository<CDAccountExecutionMetadata, String>, io.harness.repositories.executions.CDAccountExecutionMetadataRepositoryCustom {
   Optional<CDAccountExecutionMetadata> findByAccountId(String accountId);
 }
