--- 125-cd-nextgen/modules/common/core/src/main/java/io/harness/repositories/polling/BUILD.bazel
@@ -1,25 +1,0 @@
-# Copyright 2023 Harness Inc. All rights reserved.
-# Use of this source code is governed by the PolyForm Free Trial 1.0.0 license
-# that can be found in the licenses directory at the root of this repository, also available at
-# https://polyformproject.org/wp-content/uploads/2020/05/PolyForm-Free-Trial-1.0.0.txt.
-
-load("@rules_java//java:defs.bzl", "java_library")
-load("//:tools/bazel/macros.bzl", "run_analysis_per_module")
-
-java_library(
-    name = "module",
-    srcs = glob(["*.java"]),
-    visibility = ["//visibility:public"],
-    deps = [
-        "//:lombok",
-        "//125-cd-nextgen/modules/common/core/src/main/java/io/harness/polling/bean:module",
-        "//960-persistence/src/main/java/io/harness/annotation:module",
-        "//999-annotations/src/main/java/io/harness/annotations/dev:module",
-        "@maven//:com_google_inject_guice",
-        "@maven//:org_mongodb_mongo_java_driver",
-        "@maven//:org_springframework_data_spring_data_commons",
-        "@maven//:org_springframework_data_spring_data_mongodb",
-    ],
-)
-
-run_analysis_per_module()
