+++ 125-cd-nextgen/src/main/java/io/harness/polling/bean/artifact/ArtifactInfo.java
@@ -5,22 +5,21 @@
  * https://polyformproject.org/wp-content/uploads/2020/06/PolyForm-Shield-1.0.0.txt.
  */
 
-package io.harness.polling.bean;
+package io.harness.polling.bean.artifact;
 
 import io.harness.annotations.dev.HarnessTeam;
 import io.harness.annotations.dev.OwnedBy;
-import io.harness.cdng.gitpolling.bean.GitPollingConfig;
+import io.harness.cdng.artifact.bean.ArtifactConfig;
+import io.harness.delegate.task.artifacts.ArtifactSourceType;
+import io.harness.polling.bean.PollingInfo;
 
-import com.fasterxml.jackson.annotation.JsonSubTypes;
 import com.fasterxml.jackson.annotation.JsonTypeInfo;
 import com.fasterxml.jackson.annotation.JsonTypeName;
 
 @OwnedBy(HarnessTeam.CDC)
-@JsonTypeName("GITPOLLING")
+@JsonTypeName("ARTIFACT")
 @JsonTypeInfo(use = JsonTypeInfo.Id.NAME, property = "type", include = JsonTypeInfo.As.EXISTING_PROPERTY)
-@JsonSubTypes({ @JsonSubTypes.Type(value = GitHubPollingInfo.class, name = "GitHub") })
-public interface GitPollingInfo extends PollingInfo {
-  String getType();
-
-  GitPollingConfig toGitPollingConfig();
+public interface ArtifactInfo extends PollingInfo {
+  ArtifactSourceType getType();
+  ArtifactConfig toArtifactConfig();
 }
