+++ 125-cd-nextgen/src/main/java/io/harness/polling/bean/artifact/ArtifactPolledResponse.java
@@ -5,10 +5,11 @@
  * https://polyformproject.org/wp-content/uploads/2020/06/PolyForm-Shield-1.0.0.txt.
  */
 
-package io.harness.polling.bean;
+package io.harness.polling.bean.artifact;
 
 import io.harness.annotations.dev.HarnessTeam;
 import io.harness.annotations.dev.OwnedBy;
+import io.harness.polling.bean.PolledResponse;
 
 import java.util.Set;
 import lombok.Builder;
