+++ 125-cd-nextgen/src/main/java/io/harness/polling/bean/artifact/CustomArtifactInfo.java
@@ -18,7 +18,6 @@ import io.harness.cdng.artifact.bean.yaml.customartifact.CustomScriptInlineSourc
 import io.harness.cdng.artifact.bean.yaml.customartifact.FetchAllArtifacts;
 import io.harness.delegate.task.artifacts.ArtifactSourceType;
 import io.harness.pms.yaml.ParameterField;
-import io.harness.polling.bean.ArtifactInfo;
 import io.harness.steps.shellscript.ShellType;
 import io.harness.yaml.core.variables.NGVariable;
