+++ 125-cd-nextgen/src/main/java/io/harness/polling/bean/artifact/GcrArtifactInfo.java
@@ -13,7 +13,6 @@ import io.harness.cdng.artifact.bean.ArtifactConfig;
 import io.harness.cdng.artifact.bean.yaml.GcrArtifactConfig;
 import io.harness.delegate.task.artifacts.ArtifactSourceType;
 import io.harness.pms.yaml.ParameterField;
-import io.harness.polling.bean.ArtifactInfo;
 
 import lombok.Builder;
 import lombok.Value;
