+++ 125-cd-nextgen/src/main/java/io/harness/polling/bean/artifact/GoogleCloudStorageArtifactInfo.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2022 Harness Inc. All rights reserved.
+ * Copyright 2023 Harness Inc. All rights reserved.
  * Use of this source code is governed by the PolyForm Free Trial 1.0.0 license
  * that can be found in the licenses directory at the root of this repository, also available at
  * https://polyformproject.org/wp-content/uploads/2020/05/PolyForm-Free-Trial-1.0.0.txt.
@@ -10,38 +10,34 @@ package io.harness.polling.bean.artifact;
 import io.harness.annotations.dev.HarnessTeam;
 import io.harness.annotations.dev.OwnedBy;
 import io.harness.cdng.artifact.bean.ArtifactConfig;
-import io.harness.cdng.artifact.bean.yaml.AmazonS3ArtifactConfig;
-import io.harness.data.structure.EmptyPredicate;
+import io.harness.cdng.artifact.bean.yaml.GoogleCloudStorageArtifactConfig;
 import io.harness.delegate.task.artifacts.ArtifactSourceType;
 import io.harness.pms.yaml.ParameterField;
 
 import lombok.Builder;
 import lombok.Value;
 
-@OwnedBy(HarnessTeam.CDC)
+@OwnedBy(HarnessTeam.CDP)
 @Value
 @Builder
-public class S3ArtifactInfo implements ArtifactInfo {
+public class GoogleCloudStorageArtifactInfo implements ArtifactInfo {
   String connectorRef;
-  String region;
-  String bucketName;
-  String filePathRegex;
-  String filePath;
+  String project;
+  String bucket;
+  String artifactPath;
 
   @Override
   public ArtifactSourceType getType() {
-    return ArtifactSourceType.AMAZONS3;
+    return ArtifactSourceType.GOOGLE_CLOUD_STORAGE_ARTIFACT;
   }
 
   @Override
   public ArtifactConfig toArtifactConfig() {
-    return AmazonS3ArtifactConfig.builder()
+    return GoogleCloudStorageArtifactConfig.builder()
         .connectorRef(ParameterField.<String>builder().value(connectorRef).build())
-        .region(
-            ParameterField.<String>builder().value(EmptyPredicate.isNotEmpty(region) ? region : "us-east-1").build())
-        .bucketName(ParameterField.<String>builder().value(bucketName).build())
-        .filePath(ParameterField.<String>builder().value(filePath).build())
-        .filePathRegex(ParameterField.<String>builder().value(filePathRegex).build())
+        .project(ParameterField.<String>builder().value(project).build())
+        .bucket(ParameterField.<String>builder().value(bucket).build())
+        .artifactPath(ParameterField.<String>builder().value(artifactPath).build())
         .build();
   }
 }
