+++ 125-cd-nextgen/src/main/java/io/harness/polling/bean/artifact/Nexus2RegistryArtifactInfo.java
@@ -17,7 +17,6 @@ import io.harness.cdng.artifact.bean.yaml.nexusartifact.NexusRegistryNpmConfig;
 import io.harness.cdng.artifact.bean.yaml.nexusartifact.NexusRegistryNugetConfig;
 import io.harness.delegate.task.artifacts.ArtifactSourceType;
 import io.harness.pms.yaml.ParameterField;
-import io.harness.polling.bean.ArtifactInfo;
 
 import software.wings.utils.RepositoryFormat;
