+++ 125-cd-nextgen/src/main/java/io/harness/polling/bean/artifact/NexusRegistryArtifactInfo.java
@@ -19,7 +19,6 @@ import io.harness.cdng.artifact.bean.yaml.nexusartifact.NexusRegistryNugetConfig
 import io.harness.cdng.artifact.bean.yaml.nexusartifact.NexusRegistryRawConfig;
 import io.harness.delegate.task.artifacts.ArtifactSourceType;
 import io.harness.pms.yaml.ParameterField;
-import io.harness.polling.bean.ArtifactInfo;
 
 import software.wings.utils.RepositoryFormat;
