+++ 125-cd-nextgen/src/main/java/io/harness/polling/bean/gitpolling/GitPollingInfo.java
@@ -5,23 +5,23 @@
  * https://polyformproject.org/wp-content/uploads/2020/06/PolyForm-Shield-1.0.0.txt.
  */
 
-package io.harness.polling.bean;
+package io.harness.polling.bean.gitpolling;
 
 import io.harness.annotations.dev.HarnessTeam;
 import io.harness.annotations.dev.OwnedBy;
-import io.harness.cdng.manifest.yaml.ManifestOutcome;
-import io.harness.cdng.manifest.yaml.storeConfig.StoreConfig;
+import io.harness.cdng.gitpolling.bean.GitPollingConfig;
+import io.harness.polling.bean.PollingInfo;
 
 import com.fasterxml.jackson.annotation.JsonSubTypes;
 import com.fasterxml.jackson.annotation.JsonTypeInfo;
 import com.fasterxml.jackson.annotation.JsonTypeName;
 
 @OwnedBy(HarnessTeam.CDC)
-@JsonTypeName("MANIFEST")
+@JsonTypeName("GITPOLLING")
 @JsonTypeInfo(use = JsonTypeInfo.Id.NAME, property = "type", include = JsonTypeInfo.As.EXISTING_PROPERTY)
-@JsonSubTypes({ @JsonSubTypes.Type(value = HelmChartManifestInfo.class, name = "HelmChart") })
-public interface ManifestInfo extends PollingInfo {
+@JsonSubTypes({ @JsonSubTypes.Type(value = GitHubPollingInfo.class, name = "GitHub") })
+public interface GitPollingInfo extends PollingInfo {
   String getType();
-  StoreConfig getStore();
-  ManifestOutcome toManifestOutcome();
+
+  GitPollingConfig toGitPollingConfig();
 }
