+++ 125-cd-nextgen/src/main/java/io/harness/polling/bean/manifest/ManifestInfo.java
@@ -5,20 +5,24 @@
  * https://polyformproject.org/wp-content/uploads/2020/06/PolyForm-Shield-1.0.0.txt.
  */
 
-package io.harness.polling.bean;
+package io.harness.polling.bean.manifest;
 
 import io.harness.annotations.dev.HarnessTeam;
 import io.harness.annotations.dev.OwnedBy;
-import io.harness.cdng.artifact.bean.ArtifactConfig;
-import io.harness.delegate.task.artifacts.ArtifactSourceType;
+import io.harness.cdng.manifest.yaml.ManifestOutcome;
+import io.harness.cdng.manifest.yaml.storeConfig.StoreConfig;
+import io.harness.polling.bean.PollingInfo;
 
+import com.fasterxml.jackson.annotation.JsonSubTypes;
 import com.fasterxml.jackson.annotation.JsonTypeInfo;
 import com.fasterxml.jackson.annotation.JsonTypeName;
 
 @OwnedBy(HarnessTeam.CDC)
-@JsonTypeName("ARTIFACT")
+@JsonTypeName("MANIFEST")
 @JsonTypeInfo(use = JsonTypeInfo.Id.NAME, property = "type", include = JsonTypeInfo.As.EXISTING_PROPERTY)
-public interface ArtifactInfo extends PollingInfo {
-  ArtifactSourceType getType();
-  ArtifactConfig toArtifactConfig();
+@JsonSubTypes({ @JsonSubTypes.Type(value = HelmChartManifestInfo.class, name = "HelmChart") })
+public interface ManifestInfo extends PollingInfo {
+  String getType();
+  StoreConfig getStore();
+  ManifestOutcome toManifestOutcome();
 }
