+++ 125-cd-nextgen/src/main/java/io/harness/polling/mapper/manifest/GcsHelmChartManifestInfoBuilder.java
@@ -12,8 +12,8 @@ import io.harness.annotations.dev.OwnedBy;
 import io.harness.cdng.manifest.yaml.GcsStoreConfig;
 import io.harness.k8s.model.HelmVersion;
 import io.harness.pms.yaml.ParameterField;
-import io.harness.polling.bean.HelmChartManifestInfo;
 import io.harness.polling.bean.PollingInfo;
+import io.harness.polling.bean.manifest.HelmChartManifestInfo;
 import io.harness.polling.contracts.GcsHelmPayload;
 import io.harness.polling.contracts.PollingPayloadData;
 import io.harness.polling.mapper.PollingInfoBuilder;
