+++ 125-cd-nextgen/src/main/java/io/harness/polling/mapper/manifest/S3HelmChartManifestInfoBuilder.java
@@ -12,8 +12,8 @@ import io.harness.annotations.dev.OwnedBy;
 import io.harness.cdng.manifest.yaml.S3StoreConfig;
 import io.harness.k8s.model.HelmVersion;
 import io.harness.pms.yaml.ParameterField;
-import io.harness.polling.bean.HelmChartManifestInfo;
 import io.harness.polling.bean.PollingInfo;
+import io.harness.polling.bean.manifest.HelmChartManifestInfo;
 import io.harness.polling.contracts.PollingPayloadData;
 import io.harness.polling.contracts.S3HelmPayload;
 import io.harness.polling.mapper.PollingInfoBuilder;
