+++ 125-cd-nextgen/src/main/java/io/harness/polling/service/impl/PollingPerpetualTaskServiceImpl.java
@@ -21,9 +21,9 @@ import io.harness.perpetualtask.PerpetualTaskClientContextDetails;
 import io.harness.perpetualtask.PerpetualTaskExecutionBundle;
 import io.harness.perpetualtask.PerpetualTaskId;
 import io.harness.perpetualtask.PerpetualTaskSchedule;
-import io.harness.polling.bean.GitPollingInfo;
 import io.harness.polling.bean.PollingDocument;
 import io.harness.polling.bean.PollingType;
+import io.harness.polling.bean.gitpolling.GitPollingInfo;
 import io.harness.polling.service.impl.artifact.ArtifactPerpetualTaskHelperNg;
 import io.harness.polling.service.impl.gitpolling.GitPollingPerpetualTaskHelperNg;
 import io.harness.polling.service.impl.manifest.ManifestPerpetualTaskHelperNg;
