+++ 125-cd-nextgen/src/test/java/io/harness/polling/mapper/PollingDocumentMapperTest.java
@@ -21,12 +21,12 @@ import io.harness.cdng.CDNGTestBase;
 import io.harness.cdng.manifest.ManifestStoreType;
 import io.harness.cdng.manifest.yaml.GcsStoreConfig;
 import io.harness.cdng.manifest.yaml.S3StoreConfig;
-import io.harness.polling.bean.HelmChartManifestInfo;
 import io.harness.polling.bean.PollingDocument;
 import io.harness.polling.bean.artifact.AcrArtifactInfo;
 import io.harness.polling.bean.artifact.DockerHubArtifactInfo;
 import io.harness.polling.bean.artifact.EcrArtifactInfo;
 import io.harness.polling.bean.artifact.GcrArtifactInfo;
+import io.harness.polling.bean.manifest.HelmChartManifestInfo;
 import io.harness.polling.contracts.AcrPayload;
 import io.harness.polling.contracts.DockerHubPayload;
 import io.harness.polling.contracts.EcrPayload;
