+++ 125-cd-nextgen/src/test/java/io/harness/polling/service/impl/PollingPerpetualTaskServiceTest.java
@@ -51,11 +51,11 @@ import io.harness.perpetualtask.polling.ArtifactCollectionTaskParamsNg;
 import io.harness.perpetualtask.polling.GitPollingTaskParamsNg;
 import io.harness.perpetualtask.polling.ManifestCollectionTaskParamsNg;
 import io.harness.pms.yaml.ParameterField;
-import io.harness.polling.bean.GitHubPollingInfo;
-import io.harness.polling.bean.HelmChartManifestInfo;
 import io.harness.polling.bean.PollingDocument;
 import io.harness.polling.bean.PollingType;
 import io.harness.polling.bean.artifact.DockerHubArtifactInfo;
+import io.harness.polling.bean.gitpolling.GitHubPollingInfo;
+import io.harness.polling.bean.manifest.HelmChartManifestInfo;
 import io.harness.polling.service.impl.artifact.ArtifactPerpetualTaskHelperNg;
 import io.harness.polling.service.impl.gitpolling.GitPollingPerpetualTaskHelperNg;
 import io.harness.polling.service.impl.manifest.ManifestPerpetualTaskHelperNg;
