+++ 125-cd-nextgen/src/test/java/io/harness/polling/service/intfc/PollingServiceTest.java
@@ -22,12 +22,12 @@ import io.harness.annotations.dev.HarnessTeam;
 import io.harness.annotations.dev.OwnedBy;
 import io.harness.category.element.UnitTests;
 import io.harness.exception.InvalidRequestException;
-import io.harness.polling.bean.HelmChartManifestInfo;
-import io.harness.polling.bean.ManifestInfo;
 import io.harness.polling.bean.PollingDocument;
 import io.harness.polling.bean.PollingInfo;
 import io.harness.polling.bean.PollingType;
 import io.harness.polling.bean.artifact.DockerHubArtifactInfo;
+import io.harness.polling.bean.manifest.HelmChartManifestInfo;
+import io.harness.polling.bean.manifest.ManifestInfo;
 import io.harness.polling.contracts.PollingItem;
 import io.harness.polling.mapper.PollingDocumentMapper;
 import io.harness.polling.service.impl.PollingServiceImpl;
